Fixing Broken Image Build
