--- playbook/app/pb_kits/playbook/pb_lightbox/Lightbox/styles.scss
@@ -1,50 +1,0 @@
-.TagCounter{
-  display: flex;
-  justify-content: center;
-}
-
-.TagsList{
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-wrap: wrap;
-  max-width: 750px;
-  justify-content: flex-start;
-}
-
-.Hidden {
-  transform: translateY(-150px);
-}
-
-.EstimatePhotoViewer {
-  overflow-y: scroll;
-}
-
-.Header {
-  position: sticky;
-  top: 0;
-  z-index: 1;
-  background-color: #fff;
-}
-
-.PhotoViewer::before {
-  content: '';
-  width: 0;
-  padding-bottom: 100%;
-  grid-row: 1 / 1;
-  grid-column: 1 / 1;
-}
-
-.PhotoViewer > *:first-child {
-  grid-row: 1 / 1;
-  grid-column: 1 / 1;
-}
-
-.PhotoViewer {
-  display: grid;
-  padding: 2px;
-  overflow-x: hidden;
-  overflow-y: scroll;
-  grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
-  grid-auto-rows: 1fr;
-}
