+++ playbook/app/pb_kits/playbook/pb_lightbox/docs/_lightbox_multiple.jsx
@@ -3,8 +3,7 @@
 import React, { useState } from 'react'
 import { useToggler } from '../hooks/useToggler.js'
 import { Flex, Image } from '../../'
-import styles from '../Lightbox/styles.scss'
-import Lightbox from '../Lightbox/index.jsx'
+import Lightbox from '../_lightbox.jsx'
 
 export default function LightboxMultiple() {
   const photos = [
@@ -26,7 +25,7 @@ export default function LightboxMultiple() {
   }
 
   return (
-    <div className={styles.EstimatePhotoViewer}>
+    <div>
       {light ? (
         <Lightbox
             icon="times"
@@ -36,7 +35,7 @@ export default function LightboxMultiple() {
             photos={photos}
         />
       ) : (
-        <div className={styles.PhotoViewer}>
+        <div className="PhotoViewer">
           <Flex>
             {photos.map((photo, index) => {
               return (
